refactor entities

+++ src/main/kotlin/is/labs/laba_1/entities/CoordinatesEntity.kt
@@ -1,7 +1,6 @@
 package `is`.labs.laba_1.entities
 
 import jakarta.persistence.*
-import java.math.BigInteger
 
 @Entity
 @Table(name = "coordinates", schema = "is_lab_one", catalog = "postgres")
@@ -12,8 +11,8 @@ open class CoordinatesEntity (
     open var id: Int? = null,
 
     @Basic
-    @Column(name = "x", nullable = true, precision = 0)
-    open var x: BigInteger? = null,
+    @Column(name = "x", nullable = true)
+    open var x: Double? = null,
 
     @Basic
     @Column(name = "y", nullable = false)
